Extracted line chart into its own component
